src/test/java/com/javid/spring5/mvc/rest/cotrollers/v1/CustomerControllerTest.java
package com.javid.spring5.mvc.rest.cotrollers.v1;

import com.javid.spring5.mvc.rest.api.v1.model.CustomerDTO;
import com.javid.spring5.mvc.rest.services.CustomerService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.List;

import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.hasSize;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

/**
 * Created by Javid on 3/26/2021.
 */

@ExtendWith(MockitoExtension.class)
class CustomerControllerTest {

    public static final String FIRST_NAME = "Joe";
    public static final String LAST_NAME = "Dalton";
    public static final String CUSTOMER_URL = "/api/v1/customers/1";

    @Mock
    CustomerService customerService;

    @InjectMocks
    CustomerController customerController;

    MockMvc mockMvc;

    @BeforeEach
    void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(customerController).build();
    }

    @Test
    void getCustomers() throws Exception {
        // given
        var customerDTOS = List.of(new CustomerDTO().setFirstName(FIRST_NAME).setLastName(LAST_NAME),
                new CustomerDTO().setFirstName("Averell").setLastName(LAST_NAME));
        when(customerService.findAll()).thenReturn(customerDTOS);
        // when
        mockMvc.perform(get("/api/v1/customers").contentType(MediaType.APPLICATION_JSON))
                // then
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.customers", hasSize(2)));
    }

    @Test
    void getCustomer() throws Exception {
        // given
        var customerDTO = new CustomerDTO().setFirstName(FIRST_NAME)
                .setLastName(LAST_NAME).setCustomerUrl(CUSTOMER_URL);
        when(customerService.findById(anyLong())).thenReturn(customerDTO);
        // when
        mockMvc.perform(get(CUSTOMER_URL).contentType(MediaType.APPLICATION_JSON))
                // then
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.firstName", equalTo(FIRST_NAME)))
                .andExpect(jsonPath("$.lastName", equalTo(LAST_NAME)))
                .andExpect(jsonPath("$.customerUrl", equalTo(CUSTOMER_URL)));
    }
}